GH-139: Move all user-visible configuration properties to single configuration class (#140)
